backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
motor==3.7.0
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.13.0
uvicorn==0.34.0
